+++ src/components/index.ts
@@ -6,6 +6,7 @@ export { DeckGLMap } from './DeckGLMap';
 export { MapContainer, type MapView, type TimeRange, type MapContainerState } from './MapContainer';
 export * from './NewsPanel';
 export * from './MarketPanel';
+export * from './AssetIntelligencePanel';
 export * from './StockAnalysisPanel';
 export * from './StockBacktestPanel';
 export * from './PredictionPanel';